backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
neo4j==5.14.0
redis==5.0.1
sqlalchemy==2.0.23
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
